+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "preinstall": "git submodule update -i -r -f",
     "start": "nw",
-    "submodules-update": "rm -rfv node_modules/broccoli-html-editor; rm -rfv node_modules/broccoli-field-table; rm -rfv node_modules/node-px2dt-localdata-access; rm -rfv node_modules/node-express-pickles2; npm install; gulp;",
+    "submodules-update": "rm -rfv node_modules/broccoli-html-editor; rm -rfv node_modules/broccoli-field-table; rm -rfv node_modules/node-px2dt-localdata-access; rm -rfv node_modules/express-pickles2; npm install; gulp;",
     "build": "node ./build.js"
   },
   "author": "Tomoya Koyanagi",
